PoW functions: remove txs_with_pow parameter
In the iota_client_attach_to_tangle() pow_handler_t functions,
this parameter was the array to be filled by transactions with PoW
data. In order to avoid keeping in memory both the transactions
without PoW and the transactios with PoW, we now use the txs
argument to hold the list of transactions with PoW (which overwrite
the transactions without PoW that were passed when calling the
functions).